requirements.txt
Pybooru
rich
pyyaml